Replaced method of handling mixed unary and binary operators
Replaced method of handling mixed unary and binary operators.

Before, when a binary operator was hit, whether or not to pay attention to it was determined by a function peeking to either side of it by one character, paying attention to the precedences of the seen operators in different unary/binary forms.

Now, when a binary operator is hit, whether or not to pay attention to it is determined by whether all operators to the left of it before hitting a non-operator character (or the end of the string) may be postfix operators, and all operators to the right of it before hitting a non-operator character (or the end of the string) may be prefix operators.

Also, fixed:
 - % is now marked as a unary postfix operator rather than a unary prefix operator.
 - Parsing suffix operators now properly iterates through the operator group's possible suffix operators rather than prefix operators.