Rearranged default operator precedences.
Rearranged default operator precedences. Has the following changes:
- addition and subtraction binary operators now have the same precedence
- multiplication and division binary operators now have the same precedence
- positive and negative unary operators now have the same precedence
- Exponentiation now has a higher precedence than the square root unary operator.

Google calculator's rules for operator precedences were used as a reference. Modulo still has a higher precedence than multiplication and division, as it isn't invertible in the way addition/subtraction and multiplication/division are. Exponentiation is given a higher level of precedence than root as Google calculator gives it a higher level of precedence.