Added tests for custom operators
Added tests for custom operators.

fixed: Binary operators of the same precedence levels but different associativities now behave as though the left operator has the higher level of precedence. Unary operators of the same precedence levels now behave as though the prefix has the higher level of precedence. It was the reverse prior. This is somewhat arbitrary, but if forced into choosing one direction or the other, I believe binding more tightly to the left is more intuitive than binding more tightly to the right, and this is preferable to not allowing these interactions at all.