Added InvokableEvent and InvokablePriorityEvent interfaces.
Added the InvokableEvent and InvokablePriorityEvent interfaces. They inherit from Event and PriorityEvent respectively. Events that may be invoked (such as SetEvent and OrderedEvent) now inherit from these interfaces rather than from Event and PriorityEvent directly.

Interface members pertaining to the invocation of events from Event and PriorityEvent have been moved to InvokableEvent and InvokablePriorityEvent respectively. This allows Event and PriorityEvent to be exposed to the public interface of a class without offering functionality pertaining to management of the event. Previously, ProtectedEvent and ProtectedPriorityEvent were offered for blocking access to management functions of events on the public interfaces, but this solution prevents management methods from event being suggested by code completion, and allows the distinguishing of event types that do and do not allow direct invocation based on interfaces implemented.

Removed methods from ProtectedEvent and ProtectedPriorityEvent that existed purely to fulfil the Event and PriorityEvent interfaces, but were blocked.