Started working on EquationEvaluator, a rewritten version of EquationEvaluation
Created and started working on EquationEvaluator. This is to replace EquationEvaluation, implementing the builder pattern to produce an evaluator that can be re-evaluated with variables and functions swapped, but nothing deleted and operators not swapped.

This will also make use of tokens to allow for multi-character operators. (e.g. !=, <=, etc.)

So far, implemented the builder without implementing anything actually building the EquationEvaluator itself.