In EquationEvaluator.Builder, added the tracking of spaces between tokens.
In EquationEvaluator.Builder, added the tracking of spaces between tokens.

EquationEvaluator.Tokeniser now returns both a list of tokens and a list of integers, the number of spaces before the corresponding tokens.

In EquationEvaluator.Builder, parse methods now take the spacings list as well as the tokens list.

In EquationEvaluator, operator parse methods now accept the spacings list, and call builder parse methods with appropriate sublists.