In EquationEvaluator, made parse methods use TokenList
In EquationEvaluator:

Made parse methods (in Builder and in operator classes) accept and work with TokenLists rather than the lists of tokens, lists of spaces before tokens, and string representation separately.

Added and implemented methods to TokenList to accommodate this:
 - .isInBrackets()
 - .unmodifiable()
 - .withoutFirstAndLast()
 - .splitBySequence(...)

Made TokenList.splitBy(...) not split on tokens if the tokens are within brackets.