Separated variable values and function implementations of equations from their builders.
Separated variable values and function implementations of equations from their builders.

EquationComponent.evaluate(...) now requires the equation being evaluated be passed in, to derive the variable or function implementation from. A variable values or function implementation table is no longer stored on the component itself.

Partially reflected this change in the tests for Equation.Builder. Still have to reflect this change in the operator tests for Equation.Builder.