Implemented and added some tests for Equation
Implemented remaining tests for Equation.

Added and implemented the following tests for equation:
 - invalid_functionWithTrailingComma
 - invalid_functionWithLeadingComma
 - disallowed_newVariableOnEquation
 - disallowed_newFunctionOnEquation

Added "contains" methods to Equation.TokenList.

Fixed equation builders trying to parse anything following by a set of brackets as functions - if the proposed "function name" doesn't match to a function and contains operator characters, it now passes through to the nest step in parsing rather than throwing an UnrecognisedFunctionException.