In Equation, fixed inner lower-priority infix operators being parsed before nesting higher priority ones
In Equation, where a higher priority/stickier infix operator has a lower/less sticky one as a nested (not first or last) operand, fixed the inner one being parsed before the outer one.

getInfixTokenPoints now has an overload that allows it to skip over a section of the tokenlist.

tryParseOperation and tryParseInfixOperation methods now take the index of the operator priority group currently being checked.

This is passed into the new operationIsNestedInHigherPriorityInfixOperation, which checks to see if a given set of points in a given priority group is contained within a higher priority infix operation in a non-chained fashion. (that is, nested within; not the first or last operand)