Reversed the order possible tokens are passed into the tokeniser
Reversed the order possible tokens are passed into the tokeniser. This means that the tokeniser will now prefer later registered tokens to earlier registered ones.

This allows longer tokens to contain characters of shorter tokens already registered. For instance, with the default tokens registered, including the "+" token, this would allow a new token "+>" to be tokenised as expected.

This means that tokens should generally be registered in order from shortest to longest.

Added test that checks that this works as expected.