Added tests for pushing variables and functions to equations from equation builders.
Added tests for pushing variables and functions from equation builders to equations, covering the the successful case of the new variable/function implementation being used, and the case where the builder's function/variable is ignored as it's been overwritten on the equation itself.

Fixed: Builders now actually track their instances.