Added the ability to make copies of equations.
Added the ability to make copies of equations. This is done through a constructor overload accepting a source equation object. The copy has all of the variables and function implementations at the time of copying, but changes after that are not reflected.

Equations now hold a reference to the builder that's responsible for building them.

This is not made with multithreading in mind - new equation instances should be created in the lock the builder is in, even if it's as a result of copying an equation rather than building one via a builder.