Finished reïmplementing functions as mentioned in previous commit
Finished reïmplementing functions as mentioned in previous commit.

Also, swapped those functions from "forEachBranch" to "forEachItem" - there was no usage of them that didn't check if an item was present.

In Tree, added and provided a default implementation for Tree.getEntriesInOrderWhere accepting a bipredicate for the path and item, rather than a predicate for the entry.